# Test performance for calls into contracts