--- a/internal/api/lib_test.go
+++ b/internal/api/lib_test.go
@@ -8,6 +8,7 @@ import (
 	"os"
 	"path/filepath"
 	"strings"
+	"sync"
 	"testing"
 	"time"
 
@@ -610,6 +611,85 @@ func TestExecuteStorageLoop(t *testing.T) {
 	// the "sdk gas" * GasMultiplier + the wasm cost should equal the maxGas (or be very close)
 	totalCost := gasReport.UsedInternally + gasMeter2.GasConsumed()
 	require.Equal(t, int64(maxGas), int64(totalCost))
+}
+
+func BenchmarkContractCall(b *testing.B) {
+	cache, cleanup := withCache(b)
+	defer cleanup()
+
+	checksum := createCyberpunkContract(b, cache)
+
+	gasMeter1 := NewMockGasMeter(TESTING_GAS_LIMIT)
+	igasMeter1 := types.GasMeter(gasMeter1)
+	// instantiate it with this store
+	store := NewLookup(gasMeter1)
+	api := NewMockAPI()
+	querier := DefaultQuerier(MOCK_CONTRACT_ADDR, nil)
+	env := MockEnvBin(b)
+	info := MockInfoBin(b, "creator")
+
+	msg := []byte(`{}`)
+
+	res, _, err := Instantiate(cache, checksum, env, info, msg, &igasMeter1, store, api, &querier, TESTING_GAS_LIMIT, TESTING_PRINT_DEBUG)
+	require.NoError(b, err)
+	requireOkResponse(b, res, 0)
+
+	b.ResetTimer()
+	for n := 0; n < b.N; n++ {
+		gasMeter2 := NewMockGasMeter(TESTING_GAS_LIMIT)
+		igasMeter2 := types.GasMeter(gasMeter2)
+		store.SetGasMeter(gasMeter2)
+		info = MockInfoBin(b, "fred")
+		msg := []byte(`{"allocate_large_memory":{"pages":0}}`) // replace with noop once we have it
+		res, _, err = Execute(cache, checksum, env, info, msg, &igasMeter2, store, api, &querier, TESTING_GAS_LIMIT, TESTING_PRINT_DEBUG)
+		require.NoError(b, err)
+		requireOkResponse(b, res, 0)
+	}
+}
+
+func Benchmark100ConcurrentContractCalls(b *testing.B) {
+	cache, cleanup := withCache(b)
+	defer cleanup()
+
+	checksum := createCyberpunkContract(b, cache)
+
+	gasMeter1 := NewMockGasMeter(TESTING_GAS_LIMIT)
+	igasMeter1 := types.GasMeter(gasMeter1)
+	// instantiate it with this store
+	store := NewLookup(gasMeter1)
+	api := NewMockAPI()
+	querier := DefaultQuerier(MOCK_CONTRACT_ADDR, nil)
+	env := MockEnvBin(b)
+	info := MockInfoBin(b, "creator")
+
+	msg := []byte(`{}`)
+
+	res, _, err := Instantiate(cache, checksum, env, info, msg, &igasMeter1, store, api, &querier, TESTING_GAS_LIMIT, TESTING_PRINT_DEBUG)
+	require.NoError(b, err)
+	requireOkResponse(b, res, 0)
+
+	const callCount = 100 // Calls per benchmark iteration
+
+	b.ResetTimer()
+	for n := 0; n < b.N; n++ {
+		var wg sync.WaitGroup
+		wg.Add(callCount)
+		for i := 0; i < callCount; i++ {
+			go func() {
+				gasMeter2 := NewMockGasMeter(TESTING_GAS_LIMIT)
+				igasMeter2 := types.GasMeter(gasMeter2)
+				store.SetGasMeter(gasMeter2)
+				info = MockInfoBin(b, "fred")
+				msg := []byte(`{"allocate_large_memory":{"pages":0}}`) // replace with noop once we have it
+				res, _, err = Execute(cache, checksum, env, info, msg, &igasMeter2, store, api, &querier, TESTING_GAS_LIMIT, TESTING_PRINT_DEBUG)
+				require.NoError(b, err)
+				requireOkResponse(b, res, 0)
+
+				wg.Done()
+			}()
+		}
+		wg.Wait()
+	}
 }
 
 func TestExecuteUserErrorsInApiCalls(t *testing.T) {