To get an idea what it costs to call into a contract. 

In the CI and my machine this takes roughly twice the time of the "instantiate from memory" benchmark in cosmwasm-vm

| System | Instantiate from memory | Call from Go |
|---------|-------------------------|--------------|
| Circle CI docker/medium |  54 µs |   112 µs  |
| My machine  |  30 µs |  56 µs  |